requirements.txt
streamlit
torch
torchvision
torchaudio
opencv-python
numpy
Pillow
matplotlib
scikit-learn